Missing words in words file:
As can be noted in the words file, there are a couple of words that were not available in the mock database:
Words missing include: insufficiently, florey(Jack's second name), each, MIT - These words were not included.
Also important to note, in the story file, the oqzogg should be "o qzogg" as seperated.